Replace RunTests with one test exec per lib
Also run most tests without Charm++. Only run in a
Charm++ environment if needed.
