+++ cmake/SpectreAddCatchTests.cmake
@@ -49,25 +49,12 @@ list(APPEND _TEST_ENV_VARS "ASAN_OPTIONS=detect_leaks=0")
 list(APPEND _TEST_ENV_VARS "PYTHONPATH=${PYTHONPATH}")
 
 # Main function - the only one designed to be called from outside this module.
-function(spectre_add_catch_tests TEST_TARGET TEST_LIBS)
+function(spectre_add_catch_tests TEST_TARGET)
   get_target_property(SOURCE_FILES ${TEST_TARGET} SOURCES)
+  list(FILTER SOURCE_FILES EXCLUDE REGEX "TestMain.*$")
   # For each of the source files, we use spectre_parse_file to find all the
   # Catch tests inside the source file and add them to CTest.
   # We ignore the Charm++ generated header files.
-
-  # For each one of the test libraries, retrieve the source files, and
-  # the path of the library relative to ${CMAKE_SOURCE_DIR}/tests/Unit
-  # (or whatever the testing root directory is). The results are stored
-  # in "CORRECTED_LIB_SOURCES" before being added to the "SOURCE_FILES"
-  foreach (TEST_LIB ${TEST_LIBS})
-    set(CORRECTED_LIB_SOURCES "")
-    get_target_property(LIB_SOURCES ${TEST_LIB} SOURCES)
-    get_target_property(LIB_FOLDER ${TEST_LIB} FOLDER)
-    foreach (SOURCE ${LIB_SOURCES})
-      set(CORRECTED_LIB_SOURCES "${CORRECTED_LIB_SOURCES};${LIB_FOLDER}${SOURCE}")
-    endforeach()
-    set(SOURCE_FILES "${SOURCE_FILES};${CORRECTED_LIB_SOURCES}")
-  endforeach()
 
   file(MAKE_DIRECTORY "${CMAKE_BINARY_DIR}/tmp")
