+++ tests/Unit/Framework/TestingFramework.hpp
@@ -29,18 +29,6 @@
 #include "Utilities/TypeTraits/IsIterable.hpp"
 #include "Utilities/TypeTraits/IsMaplike.hpp"
 
-/// \cond
-// The macro SPECTRE_TEST_REGISTER_FUNCTION is defined inside the
-// add_test_library CMake function. It is used to make a call into a translation
-// unit so that static variables for Catch are properly initialized.
-#ifdef SPECTRE_TEST_REGISTER_FUNCTION
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wmissing-declarations"
-void SPECTRE_TEST_REGISTER_FUNCTION() {}  // NOLINT
-#pragma GCC diagnostic pop
-#endif  // SPECTRE_TEST_REGISTER_FUNCTION
-/// \endcond
-
 /*!
  * \ingroup TestingFrameworkGroup
  * \brief A replacement for Catch's TEST_CASE that silences clang-tidy warnings
