--- tests/Unit/RunTests.ci
@@ -1,10 +1,0 @@
-// Distributed under the MIT License.
-// See LICENSE.txt for details.
-
-mainmodule RunTests {
-
-  mainchare RunTests {
-    entry RunTests(CkArgMsg * msg);
-  };
-
-};
