+++ tests/Unit/TestMainCharm.ci
@@ -1,0 +1,10 @@
+// Distributed under the MIT License.
+// See LICENSE.txt for details.
+
+mainmodule TestMainCharm {
+
+  mainchare TestMainCharm {
+    entry TestMainCharm(CkArgMsg * msg);
+  };
+
+};
