+++ tests/Unit/TestMainCharm.cpp
@@ -1,13 +1,14 @@
 // Distributed under the MIT License.
 // See LICENSE.txt for details.
 
-#include "RunTests.hpp"
+#include "TestMainCharm.hpp"
 
 #include "Framework/TestingFramework.hpp"
 
 #include <charm++.h>
 #include <cstddef>
 #include <exception>
+#include <limits>
 #include <memory>
 #include <string>
 
@@ -17,11 +18,13 @@
 #include "Parallel/Printf.hpp"
 #include "Utilities/ErrorHandling/FloatingPointExceptions.hpp"
 #include "Utilities/ErrorHandling/SegfaultHandler.hpp"
+#include "Utilities/MemoryHelpers.hpp"
 #include "Utilities/System/Abort.hpp"
 #include "Utilities/System/Exit.hpp"
 
-RunTests::RunTests(CkArgMsg* msg) {
+TestMainCharm::TestMainCharm(CkArgMsg* msg) {
   setup_error_handling();
+  setup_memory_allocation_failure_reporting();
   Parallel::printf("%s", info_from_build().c_str());
   enable_floating_point_exceptions();
   enable_segfault_handler();
@@ -30,10 +33,8 @@ RunTests::RunTests(CkArgMsg* msg) {
   Catch::StringMaker<float>::precision =
       std::numeric_limits<float>::max_digits10;
   const int result = Catch::Session().run(msg->argc, msg->argv);
-  // In the case where we run all the non-failure tests at once we must ensure
-  // that we only initialize and finalize the python env once. Initialization is
-  // done in the constructor of SetupLocalPythonEnvironment, while finalization
-  // is done in the constructor of RunTests.
+  // Clean up the Python environment only after all tests have finished running,
+  // since there could be multiple tests run in a single executable launch.
   pypp::SetupLocalPythonEnvironment::finalize_env();
   if (0 == result) {
     sys::exit();
@@ -41,15 +42,4 @@ RunTests::RunTests(CkArgMsg* msg) {
   sys::abort("A catch test has failed.");
 }
 
-#include "tests/Unit/RunTests.def.h"
-
-// Needed for tests that use the GlobalCache since it registers itself with
-// Charm++. However, since Parallel/CharmMain.tpp isn't included in the RunTests
-// executable, no actual registration is done, the GlobalCache is only
-// queued for registration.
-namespace Parallel::charmxx {
-class RegistrationHelper;
-std::unique_ptr<RegistrationHelper>* charm_register_list = nullptr;
-size_t charm_register_list_capacity = 0;
-size_t charm_register_list_size = 0;
-}  // namespace Parallel::charmxx
+#include "tests/Unit/TestMainCharm.def.h"
