+++ tests/Unit/TestMainCharm.hpp
@@ -1,0 +1,16 @@
+// Distributed under the MIT License.
+// See LICENSE.txt for details.
+
+#pragma once
+
+#include "tests/Unit/TestMainCharm.decl.h"
+
+/// \cond
+class CkArgMsg;
+/// \endcond
+
+/// Main executable for running unit tests in a Charm++ environment
+class TestMainCharm : public CBase_TestMainCharm {
+ public:
+  [[noreturn]] explicit TestMainCharm(CkArgMsg* msg);
+};
